KphpErrorContext -> JsonLogger
